Updated publish scripts [skip ci].

+++ scripts/publish.sh
@@ -11,7 +11,15 @@ fi
 cargo test -- --test-threads=1
 
 for CRATE in ${CRATES[@]}; do
-  pushd $CRATE
+  echo "INFO : processing crate ${CRATE}."
+  pushd $CRATE > /dev/null
+
+  VERSION="$(cargo pkgid | cut -d# -f2)"
+
+  if ! curl -s -o /dev/null -w '%{http_code}' https://github.com/apognu/knox/tree/v${VERSION} | grep 200 > /dev/null; then
+    echo "ERROR: tag v${VERSION} does not exist in upstream git repository."
+    exit 1
+  fi
   
   cp Cargo.toml Cargo.toml.bak
   sed -i 's/build =.*/build = false/' Cargo.toml
@@ -22,5 +30,5 @@ for CRATE in ${CRATES[@]}; do
 
   mv Cargo.toml.bak Cargo.toml
   
-  popd
+  popd > /dev/null
 done
